requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
appdirs==1.4.4
autopep8==1.5.5
black==20.8b1
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fake-bpy-module-2.83==20210304
flake8==3.8.4
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.1
pathspec==0.8.1
pycodestyle==2.6.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyflakes==2.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
regex==2020.11.13
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
typed-ast==1.4.2
typing-extensions==3.7.4.3
